pyproject.toml
[tool.poetry]
name = "git-taxbreak"
version = "0.1.0"
description = "Tool for collect artifacts from git for register creative work"
authors = [
    "Luczak Kamil <kamilluczak@luczakweb.com>",
]

license = "MIT"
readme = "README.rst"

repository = "https://github.com/kamil1b/git-taxbreak"
homepage = "https://github.com/kamil1b/git-taxbreak"

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX",
    # "Operating System :: Microsoft :: Windows"
    # "Operating System :: MacOS :: MacOS X"
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
]
keywords = ["git", "taxbreak", "tool"]

[tool.poetry.dependencies]
python = ">=3.5"
gitpython = "^3"
pre-commit = "^1.20"

[tool.poetry.dev-dependencies]
pytest = ">=5.0"
pytest-cov = ">=2.6.1"
black = { version = "^19.10-beta.0", python = ">=3.6" }

[tool.poetry.scripts]
git-taxbreak = 'git_taxbreak.cmd:main'

[build-system]
requires = ["poetry>=1.0.0b3"]
build-backend = "poetry.masonry.api"

[tool.towncrier]
package = "git_taxbreak"
package_dir = "src/"
filename = "CHANGELOG.rst"
directory = "changelog/"
title_format = "git-taxbreak {version} ({project_date})"
template = "changelog/_templates.rst"

  [[tool.towncrier.type]]
  directory = "removal"
  name = "Removals"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "deprecation"
  name = "Deprecations"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "feature"
  name = "Features"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "improvement"
  name = "Improvements"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "bugfix"
  name = "Bug Fixes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "vendor"
  name = "Vendored Libraries"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "doc"
  name = "Improved Documentation"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "trivial"
  name = "Trivial/Internal Changes"
  showcontent = true

[config]
known_third_party = ["git", "git_taxbreak", "pkg_resources", "pytest"]
